+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## v0.3.19 - Dominating Set Problems 
+  * **Commit**: 2026-01-31 18:44
+  * problem.NewDominatingSet 
 ## v0.3.18 - Graph Cover Problems
   * **Commit**: 2026-01-31 18:21
   * problem.NewVertexCover